Refactor rating component
