+++ src/screens/Technology/components/About/styles.ts
