+++ src/screens/Technology/components/Rating/index.tsx
@@ -2,16 +2,55 @@ import React from 'react';
 import * as S from './styles';
 import { Tabs } from '../../../../components';
 
+interface RateCommentsProps {
+  content: string
+  author: string
+}
+interface RateCommentProps {
+  data: RateCommentsProps[]
+}
+
+const RateComments = ({ data }: RateCommentProps): JSX.Element => (
+  data && data[0] && (
+    <>
+      <S.Stars>
+        <S.Star solid name="star" size={12} color="#f9d142" />
+        <S.Star solid name="star" size={12} color="#f9d142" />
+        <S.Star solid name="star" size={12} color="#f9d142" />
+        <S.Star solid name="star" size={12} color="#f9d142" />
+        <S.Star solid name="star" size={12} color="#e8e8e8" />
+      </S.Stars>
+      <S.RateCommentAuthor>{data[0].author}</S.RateCommentAuthor>
+      <S.RateCommentText>{data[0].content}</S.RateCommentText>
+      <S.RateViewMore activeOpacity={0.7}>
+        <S.RateViewMoreText>Ver todas as opiniões</S.RateViewMoreText>
+      </S.RateViewMore>
+    </>
+  )
+);
+
 const Rating = () => (
   <S.Wrapper>
     <S.Title>Avaliações</S.Title>
-    <S.Rate />
+    <S.WrapperRate>
+      <S.RateNumber>4.7</S.RateNumber>
+      <S.RateStars>
+        <S.Stars>
+          <S.Star solid name="star" size={24} color="#f9d142" />
+          <S.Star solid name="star" size={24} color="#f9d142" />
+          <S.Star solid name="star" size={24} color="#f9d142" />
+          <S.Star solid name="star" size={24} color="#f9d142" />
+          <S.Star solid name="star" size={24} color="#e8e8e8" />
+        </S.Stars>
+        <S.RateStarsDescription>Média entre 32 opiniões</S.RateStarsDescription>
+      </S.RateStars>
+    </S.WrapperRate>
     <Tabs
       tabs={[
         {
           title: 'Todos',
           content: (
-            <S.RateComments
+            <RateComments
               data={[
                 {
                   content: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
@@ -24,7 +63,7 @@ const Rating = () => (
         {
           title: 'Positivos',
           content: (
-            <S.RateComments
+            <RateComments
               data={[
                 {
                   content: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
@@ -37,7 +76,7 @@ const Rating = () => (
         {
           title: 'Negativos',
           content: (
-            <S.RateComments
+            <RateComments
               data={[
                 {
                   content: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
